+++ src/hash/LSH.h
@@ -33,28 +33,28 @@ class LSH {
 
   static std::vector<int64_t> getMinHash(
     const std::string& seq, 
-    int K, 
-    int W,
+    size_t K, 
+    size_t W,
     HashType hashType = HashType::MD5, 
     int MOD = 0
   );
 
   static std::vector<int64_t> getMaxHash(
     const std::string& seq, 
-    int K, 
-    int W,
+    size_t K, 
+    size_t W,
     HashType hashType = HashType::MD5, 
     int MOD = 0
   );
 
   static int64_t getKMerValuePlain(
-    const std::string& seq, int64_t offset, int64_t K);
+    const std::string& seq, size_t offset, size_t K);
 
   static int64_t getKMerValuePlainMOD(
-    const std::string& seq, int64_t offset, int64_t K, int64_t MOD);
+    const std::string& seq, size_t offset, size_t K, int64_t MOD);
 
   static int64_t getKMerValueMD5(
-    const std::string& seq, int64_t offset, int64_t K);
+    const std::string& seq, size_t offset, size_t K);
 
   static int64_t md5ToInt(const char* fingerprint);
